Update styles.css added Certification section, repositioning Social icon, update upcoming projects and update the images.
Added the certification Section for showcase my certificates. Adjusted and and repositioning the social media icon and contacts. Change in the project section, update the upcoming projects, add Medicine Recommendation System as upcoming project. Giving them styles and colours. Also change and update the images and documents, which are required. And some minor changes ...